+++ geo/CHANGES.md
@@ -13,6 +13,8 @@
   <https://github.com/georust/geo/pull/987>
 - Added `MinimumRotatedRect` trait to calculate the MBR of geometry
   <https://github.com/georust/geo/pull/959>
+- Added `GeodesicArea` trait to support geodesic area and perimeter calculations from `geographlib-rs`
+  <https://github.com/georust/geo/pull/988>
 
 ## 0.23.1